requirements.txt
Click==7.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.10.3
